requirements.txt
discord.py
requests
python-dotenv
discord-py-interactions
asyncio